LayoutTests/imported/w3c/web-platform-tests/css/css-values/clamp-length-computed-expected.txt


FAIL Property letter-spacing value 'clamp(10px, 20px, 30px)' computes to '20px' assert_equals: expected "20px" but got "normal"
FAIL Property letter-spacing value 'clamp(10px, 5px, 30px)' computes to '10px' assert_equals: expected "10px" but got "normal"
FAIL Property letter-spacing value 'clamp(10px, 35px, 30px)' computes to '30px' assert_equals: expected "30px" but got "normal"
FAIL Property letter-spacing value 'clamp(30px, 100px, 20px)' computes to '30px' assert_equals: expected "30px" but got "normal"
